Initial support for using API size values for single photos

+++ app/src/main/java/com/h/pixeldroid/utils/api/objects/Attachment.kt
@@ -28,8 +28,8 @@ data class Attachment(
 
     {
         data class Focus(
-                val x: Int?,
-                val y: Int?
+                val x: Double?,
+                val y: Double?
         ) : Serializable
         data class Image(
                 val width: Int?,
